+++ maplibre_gl_platform_interface/lib/src/method_channel_maplibre_gl.dart
@@ -288,10 +288,13 @@ class MapLibreMethodChannel extends MapLibrePlatform {
   Future<bool> easeCamera(
     CameraUpdate cameraUpdate, {
     Duration? duration,
+    CameraAnimationInterpolation? interpolation,
   }) async {
     final result = await _channel.invokeMethod('camera#ease', <String, dynamic>{
       'cameraUpdate': cameraUpdate.toJson(),
       'duration': duration?.inMilliseconds,
+      if (interpolation != null)
+        'interpolation': interpolation.toString().split('.').last,
     });
     return result as bool;
   }